# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Site da SIFSoft jr 2.0
  description: Documentação do sistema de gerenciamento.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: API SIFSoft
paths:
...
  /login:
    post:
      tags:
        - Login
      summary: Realizar login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginSchema'
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Credenciais inválidas
  /usuarios:
    post:
      tags:
        - Usuários
      summary: Cria um novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserSchema'
      responses:
        '201':
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Dados inválidos
    get:
      tags:
        - Usuários
      summary: Lista todos os usuários (admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          description: Acesso negado
...
  /postagens:
    post:
      tags:
        - Postagens
      summary: Cria uma nova postagem (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PostagemCreateSchema'
      responses:
        '201':
          description: Postagem criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Postagem'
        '403':
          description: Acesso negado
    get:
      tags:
        - Postagens
      summary: Lista postagens com paginação
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Lista de postagens
          content:
            application/json:
              schema:
                type: object
                properties:
                  postagens:
                    type: array
                    items:
                      $ref: '#/components/schemas/Postagem'
                  total:
                    type: integer
                  page:
                    type: integer
                  pages:
                    type: integer
...
  /comentarios:
    post:
      tags:
        - Comentários
      summary: Cria um novo comentário (autenticado)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComentarioCreateSchema'
      responses:
        '201':
          description: Comentário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comentario'
        '401':
          description: Não autenticado
...
  /editais:
    get:
      tags:
        - Editais
      summary: Lista todos os editais
      responses:
        '200':
          description: Lista de editais
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Edital'
    post:
      tags:
        - Editais
      summary: Cria um novo edital (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEditalSchema'
      responses:
        '201':
          description: Edital criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edital'
        '403':
          description: Acesso negado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateUserSchema:
      type: object
      required:
        - nome
        - email
        - senha
      properties:
        nome:
          type: string
          minLength: 3
          maxLength: 100
        email:
          type: string
          format: email
        senha:
          type: string
          minLength: 6
          maxLength: 100
...
    User:
      type: object
      properties:
        idUsuario:
          type: integer
        nome:
          type: string
        email:
          type: string
        perfilId:
          type: integer
        perfil:
          $ref: '#/components/schemas/Perfil'
    Perfil:
      type: object
      properties:
        idPerfil:
          type: integer
        nome_perfil:
          type: string
    LoginSchema:
      type: object
      required:
        - email
        - senha
      properties:
        email:
          type: string
          format: email
        senha:
          type: string
          minLength: 6
          maxLength: 100
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            usuario:
              $ref: '#/components/schemas/User'
            token:
              type: string
    PostagemCreateSchema:
      type: object
      required:
        - titulo
        - conteudo
        - resumo
        - tipo
      properties:
        titulo:
          type: string
          minLength: 3
          maxLength: 255
        conteudo:
          type: string
          minLength: 10
          maxLength: 50000
        resumo:
          type: string
          minLength: 10
          maxLength: 500
        tipo:
          type: string
          enum: [post, servico]
        imagem_capa:
          type: string
          format: binary
    Postagem:
      type: object
      properties:
        idPostagem:
          type: integer
        titulo:
          type: string
        conteudo:
          type: string
        resumo:
          type: string
        imagem_capa:
          type: string
        tipo:
          type: string
        data:
          type: string
          format: date-time
        autorId:
          type: integer
        autor:
          $ref: '#/components/schemas/User'
    ComentarioCreateSchema:
      type: object
      required:
        - texto
        - usuarioId
        - postagemId
      properties:
        texto:
          type: string
          minLength: 1
          maxLength: 2000
        usuarioId:
          type: integer
        postagemId:
          type: integer
...
    Comentario:
      type: object
      properties:
        idComentario:
          type: integer
        texto:
          type: string
        data:
          type: string
          format: date-time
        usuarioId:
          type: integer
        usuario:
          $ref: '#/components/schemas/User'
        postagemId:
          type: integer
        postagem:
          $ref: '#/components/schemas/Postagem'
    CreateEditalSchema:
      type: object
      required:
        - titulo
        - tipo
        - editalurl
        - data_validade
        - autorId
      properties:
        titulo:
          type: string
          minLength: 1
        tipo:
          type: string
          minLength: 1
        editalurl:
          type: string
          minLength: 1
        data_publicacao:
          type: string
          format: date-time
        data_validade:
          type: string
          format: date-time
        autorId:
          type: integer
...
    Edital:
      type: object
      properties:
        idEdital:
          type: integer
        titulo:
          type: string
        tipo:
          type: string
        editalurl:
          type: string
        data_publicacao:
          type: string
          format: date-time
        data_validade:
          type: string
          format: date-time
        autorId:
          type: integer
        autor:
          $ref: '#/components/schemas/User'